put method call option separate
